feat : 이벤트 매니저 추가, movement 컴포넌트 구성, 애니메이션 보간 정상화

+++ EngineSDK/inc/GameObject.h
@@ -91,7 +91,7 @@ public:
 public:
      template <typename T> requires is_base_of_v<Component, T>
       	Shared<T> Get_Component() {
-        uint32 typeID = rttr::type::get<T>().get_id();
+        uint32 typeID = static_cast<uint32>(rttr::type::get<T>().get_id());
 
         for (auto &[objectID, component] : m_Components) {
             if (component->Get_TypeID() == typeID) {
